update README and add conditional rendering
