+++ readme.md
@@ -1,46 +1,173 @@
-# Getting Started with Create React App and Redux
+<a name="readme-top"></a>
+<div align="center">
 
-This project was bootstrapped with [Create React App](https://github.com/facebook/create-react-app), using the [Redux](https://redux.js.org/) and [Redux Toolkit](https://redux-toolkit.js.org/) template.
+  <img src="murple_logo.png" alt="logo" width="140"  height="auto" />
+  <br/>
 
-## Available Scripts
+<h3><b>Microverse README Template</b></h3>
 
-In the project directory, you can run:
+</div>
 
-### `npm start`
+<!-- TABLE OF CONTENTS -->
 
-Runs the app in the development mode.\
-Open [http://localhost:3000](http://localhost:3000) to view it in your browser.
+# 📗 Table of Contents
 
-The page will reload when you make changes.\
-You may also see any lint errors in the console.
+- [📖 About the Project](#about-project)
+    - [🛠 Built With](#built-with)
+        - [Tech Stack](#tech-stack)
+        - [Key Features](#key-features)
+    - [🚀 Live Demo](#live-demo)
+- [💻 Getting Started](#getting-started)
+    - [Setup](#setup)
+    - [Prerequisites](#prerequisites)
+    - [Install](#install)
+    - [Usage](#usage)
+    - [Run tests](#run-tests)
+    - [Deployment](#triangular_flag_on_post-deployment)
+- [👥 Authors](#authors)
+- [🔭 Future Features](#future-features)
+- [🤝 Contributing](#contributing)
+- [⭐️ Show your support](#support)
+- [🙏 Acknowledgements](#acknowledgements)
+- [❓ FAQ](#faq)
+- [📝 License](#license)
 
-### `npm test`
+<!-- PROJECT DESCRIPTION -->
 
-Launches the test runner in the interactive watch mode.\
-See the section about [running tests](https://facebook.github.io/create-react-app/docs/running-tests) for more information.
+# 📖 Jetlog Front-End <a name="about-project"></a>
 
-### `npm run build`
+This is a real world practical app for renting planes. When the user visits the app they can see all available planes and click on a specific plane to see more details. The user is also authenticated to reserve a plane and see their specific reservations. The app has additional authorization for the admin to add or delete a plane from the database. It's build alongside the [jetlog back-end application](https://github.com/tkz96/jet-log-backend)
 
-Builds the app for production to the `build` folder.\
-It correctly bundles React in production mode and optimizes the build for the best performance.
+## 🛠 Built With <a name="built-with"></a>
 
-The build is minified and the filenames include the hashes.\
-Your app is ready to be deployed!
+### Tech Stack <a name="tech-stack"></a>
 
-See the section about [deployment](https://facebook.github.io/create-react-app/docs/deployment) for more information.
 
-### `npm run eject`
+<details>
+  <summary>Client</summary>
+  <ul>
+    <li><a href="https://reactjs.org/">React</a></li>
+    <li><a href="https://redux.js.org/">Redux</a></li>
+  </ul>
+</details>
 
-**Note: this is a one-way operation. Once you `eject`, you can't go back!**
+<details>
+  <summary>Server</summary>
+  <ul>
+    <li><a href="https://render.com/">Render</a></li>
+  </ul>
+</details>
 
-If you aren't satisfied with the build tool and configuration choices, you can `eject` at any time. This command will remove the single build dependency from your project.
+<details>
+<summary>Database</summary>
+  <ul>
+    <li><a href="https://www.postgresql.org/">PostgreSQL</a></li>
+  </ul>
+</details>
 
-Instead, it will copy all the configuration files and the transitive dependencies (webpack, Babel, ESLint, etc) right into your project so you have full control over them. All of the commands except `eject` will still work, but they will point to the copied scripts so you can tweak them. At this point you're on your own.
+<p align="right">(<a href="#readme-top">back to top</a>)</p>
 
-You don't have to ever use `eject`. The curated feature set is suitable for small and middle deployments, and you shouldn't feel obligated to use this feature. However we understand that this tool wouldn't be useful if you couldn't customize it when you are ready for it.
+<!-- GETTING STARTED -->
 
-## Learn More
+## 💻 Getting Started <a name="getting-started"></a>
 
-You can learn more in the [Create React App documentation](https://facebook.github.io/create-react-app/docs/getting-started).
+To get a local copy up and running, follow these steps.
 
-To learn React, check out the [React documentation](https://reactjs.org/).
+### Prerequisites
+
+In order to run this project you need:
+- <a href="https://github.com/tkz96/jet-log-backend">Jetlog Backend</a>
+- Node and NPM installed
+- git installed [In case you want to clone this repo using your terminal]
+- Any IDE or text editor that supports Javascript
+
+### Setup
+
+Clone this repository to your desired folder:
+
+#### 1. Clone this repository:
+
+```bash command
+git clone https://github.com/tkz96/jet-log-frontend.git
+```
+
+#### 2. Navigate to the location of the folder in your machine:
+
+```bash command
+cd jet-log-frontend
+```
+
+#### 3. Install the node packages:
+
+```bash command
+npm install
+```
+
+#### 4. Finally you can run the server:
+
+```bash command
+npm start
+```
+
+#### 5. To run tests:
+
+```bash command
+npm test
+```
+
+<p align="right">(<a href="#readme-top">back to top</a>)</p>
+
+<!-- AUTHORS -->
+
+## 👥 Authors <a name="authors"></a>
+
+👤 **Talha Zuberi**
+
+- GitHub: [@tkz96](https://github.com/tkz96)
+- Twitter: [@tkz96](https://twitter.com/tkz96)
+- LinkedIn: [Talha Zuberi](https://www.linkedin.com/in/talha-zuberi/)
+
+👤 **Bengi Genç**
+
+- GitHub: [@bengigo](https://github.com/bengigo)
+- Twitter: [@bengigo](https://twitter.com/bengi_gb)
+- LinkedIn: [Bengi](https://www.linkedin.com/in/bengigenc/)
+
+
+👤 **Sharon Odhiambo**
+
+- GitHub: [@sharon-odhiambo](https://github.com/sharon-odhiambo)
+- Twitter: [@SharonVictor16](https://twitter.com/sharonvictor16)
+- LinkedIn: [Sharon Odhiambo](https://www.linkedin.com/in/sharonn-odhiambo/)
+
+👤 **Aquinas Moriasi**
+
+- GitHub: [Aquinas Moriasi](https://github.com/Aquinasmoriasi)
+- Twitter: [@Aquinas747](https://twitter.com/Aquinas747)
+- LinkedIn: [Aquinas Moriasi](https://twitter.com/aquinas-moriasi)
+
+👤 **Berkay Akbas**
+
+- GitHub: [@berkay-akbas](https://github.com/Berkay-akbas)
+- LinkedIn: [Berkay Akbas](https://www.linkedin.com/in/berkay-akbas-a03b3b239/)
+
+
+<p align="right">(<a href="#readme-top">back to top</a>)</p>
+
+<!-- CONTRIBUTING -->
+
+## 🤝 Contributing <a name="contributing"></a>
+
+Contributions, issues, and feature requests are welcome!
+
+Feel free to check the [issues page](../../issues/).
+
+<p align="right">(<a href="#readme-top">back to top</a>)</p>
+
+<!-- SUPPORT -->
+
+## ⭐️ Show your support <a name="support"></a>⭐
+
+If you like this project, please give it a ⭐ at [jetlog-frontend](https://github.com/tkz96/jet-log-frontend) and [jetlog-backend](https://github.com/tkz96/jet-log-backend)
+
+<p align="right">(<a href="#readme-top">back to top</a>)</p>
