+++ src/components/Aeroplanes/Aeroplanes.js
@@ -1,28 +1,32 @@
 import React, { useEffect } from 'react';
 import { useDispatch } from 'react-redux';
 import { useSelector } from 'react-redux/es/exports';
+import { NavLink } from 'react-router-dom';
 import Carousel, { slidesToShowPlugin } from '@brainhubeu/react-carousel';
 import '@brainhubeu/react-carousel/lib/style.css';
+import { fetchAeroplane } from '../../redux/aeroplanes/aeroplanes';
 import '../../styles/Homepage.css';
-import { fetchAeroplane, deletePlane } from '../../redux/aeroplanes/aeroplanes';
 
-const DeleteAeroplanes = () => {
+const Aeroplanes = () => {
   const dispatch = useDispatch();
   const aeroplanes = useSelector((state) => state.aeroplanes);
   useEffect(() => {
     dispatch(fetchAeroplane());
   }, [dispatch]);
 
-  const onClickDelete = (e) => {
-    dispatch(deletePlane(e.target.id));
+  const onClickHandler = (e) => {
+    const aero = aeroplanes.find((a) => a.id === e.target.id);
+    localStorage.setItem('aero', JSON.stringify(aero));
   };
 
+  const handleDragStart = (e) => e.preventDefault();
   return (
     <>
       {!aeroplanes.length && <h6 className="text-center mt-5">No aeroplanes at the moment</h6>}
       {aeroplanes.length > 0 && (
       <div className="d-flex flex-column gap-3 justify-content-center align-items-center home">
-        <h2 className="mt-5 pt-5">Delete a Plane</h2>
+        <h2 className="mt-5 pt-5">Latest Models</h2>
+        <p className="top">Kindly select a plane for renting</p>
         <Carousel
           plugins={[
             'arrows',
@@ -63,15 +67,23 @@ const DeleteAeroplanes = () => {
         >
           {aeroplanes.map((a) => (
             <div key="aero" className="d-flex flex-column justify-content-center align-items-center gap-3">
-              <div className="back">
-                <div
-                  style={{ backgroundImage: `url(${a.image})` }}
-                  id={a.id}
-                  className="image"
-                />
-              </div>
+              <NavLink
+                key={a.links}
+                to={`/Aeroplanes/${a.id}`}
+              >
+                <div className="back">
+                  <div
+                    style={{ backgroundImage: `url(${a.image})` }}
+                    id={a.id}
+                    onClick={onClickHandler}
+                    onDragStart={handleDragStart}
+                    role="presentation"
+                    className="image"
+                  />
+                </div>
+              </NavLink>
               <span className="name pt-3">{a.name}</span>
-              <button type="button" id={a.id} onClick={onClickDelete}><span>Delete</span></button>
+              <p className="d-flex align-self-center ps-5">{a.description}</p>
             </div>
           ))}
         </Carousel>
@@ -81,4 +93,4 @@ const DeleteAeroplanes = () => {
   );
 };
 
-export default DeleteAeroplanes;
+export default Aeroplanes;
